Remove environment config
Environment was added to identify the instance of the application,
as well as to make decisions based on the instance.

Experience showed that identifing the instance can be done in the
operational context and that configuration should be used
to make those decisions.

The environment also helped to add the instance ID to logs as
structured context, but that information should come from a log
collector, enriching log events with operational metadata.